Socket recv/send support state return for async.

+++ raster/net/Socket.h
@@ -48,8 +48,16 @@ public:
   void close();
   bool isClosed();
 
-  int recv(void* buf, size_t n);
-  int send(void* buf, size_t n);
+  /*
+   * return:
+   *  >0: read/write size
+   *   0: peer is closed (make sure data exists by user)
+   *  -1: error
+   *  -2: again (nonblock) / timeout (block)
+   *  -3: peer is closed
+   */
+  ssize_t recv(void* buf, size_t n);
+  ssize_t send(void* buf, size_t n);
 
   bool setRecvTimeout(uint64_t t);
   bool setSendTimeout(uint64_t t);
